[SDAN-428] LHS navbar should only contain icon if the user has access to section.
